+++ src/plugins/point_in_time_management/public/components/utils.ts
@@ -1,0 +1,30 @@
+/*
+ * Copyright OpenSearch Contributors
+ * SPDX-License-Identifier: Apache-2.0
+ */
+
+import { SavedObjectsClientContract } from 'src/core/public';
+import { DataSourceAttributes } from 'src/plugins/data_source/common/data_sources';
+
+export async function getDataSources(savedObjectsClient: SavedObjectsClientContract) {
+  return savedObjectsClient
+    .find<DataSourceAttributes>({
+      type: 'data-source',
+      fields: ['title', 'id'],
+      perPage: 10000,
+    })
+    .then((response) => {
+      return (
+        response.savedObjects.map((dataSource) => {
+          const id = dataSource.id;
+          const title = dataSource.get('title');
+
+          return {
+            id,
+            title,
+            sort: `${title}`,
+          };
+        }) || []
+      );
+    });
+}